src/main/java/com/example/security/domain/dto/AdminDto.java
package com.example.security.domain.dto;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public class AdminDto {
    @NotEmpty(message = "Username required")
    private String username;

    @NotEmpty(message = "Password required")
    @Size(min = 8, message = "Must be at least 8 characters")
    @Pattern(regexp = "(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*", message = passwordRegexError)
    private String password;

    private String confirmationPassword;

    public static final String passwordRegexError = "Password must contain at least 1 uppercase letter and 1 number";

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getConfirmationPassword() {
        return confirmationPassword;
    }

    public void setConfirmationPassword(String confirmationPassword) {
        this.confirmationPassword = confirmationPassword;
    }

}
